+++ packages/sdk/test/unit/Resends.test.ts
@@ -8,6 +8,7 @@ import { Resends } from '../../src/subscribe/Resends'
 import { MOCK_CONTENT, mockLoggerFactory } from '../test-utils/utils'
 import { MessageID } from './../../src/protocol/MessageID'
 import { ContentType, EncryptionType, SignatureType, StreamMessage } from './../../src/protocol/StreamMessage'
+import { randomBytes } from 'crypto'
 
 const createResends = (serverUrl: string) => {
     return new Resends(
@@ -59,7 +60,7 @@ describe('Resends', () => {
         const MESSAGE_COUNT = 257
         const streamPartId = StreamPartIDUtils.parse('stream#0')
         const server = await startTestServer('/streams/:streamId/data/partitions/:partition/:resendType', async (_req, res) => {
-            const publisherId = randomEthereumAddress()
+            const publisherId = randomBytes(30)
             for (const _ of range(MESSAGE_COUNT)) {
                 const msg = new StreamMessage({
                     messageId: new MessageID(toStreamID('streamId'), 0, 0, 0, publisherId, ''),