+++ packages/sdk/test/unit/validateStreamMessage.test.ts
@@ -38,8 +38,8 @@ const validate = async (messageOptions: MessageOptions) => {
                 partitions: PARTITION_COUNT
             })
         } as any),
-        isStreamPublisher: async (_streamIdOrPath: string, userAddress: EthereumAddress) => {
-            return userAddress === toEthereumAddress(publisherWallet.address)
+        isStreamPublisher: async (_streamIdOrPath: string, userId: Uint8Array) => {
+            return areEqualBinaries(userId, hexToBinary(publisherWallet.address))
         }
     }
     await validateStreamMessage(msg, streamRegistry as any, new SignatureValidator(mock<ERC1271ContractFacade>()))