+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changes before Tatum release are not documented in this file.
 
 #### Removed
 
+- Remove support for legacy encryption keys
+
 #### Fixed
 
 #### Security